Ajoute l'activité mixte aux simulateurs indépendants
Factorise la logique de la saisie des CA dans un nouveau composant
