+++ modele-social/règles/entreprise-établissement.yaml
@@ -833,18 +833,19 @@ entreprise . activité . mixte . proportions:
     - nom: service BIC
       variations:
         - si: activité = 'libérale'
-          alors: 0
+          alors: 0%
         - sinon: 50%
     - nom: service BNC
       variations:
         - si: activité = 'libérale'
           alors: 2 / 3
-        - sinon: 0
+        - sinon: 0%
     - nom: vente restauration hébergement
       variations:
         - si: activité = 'libérale'
           alors: 1 / 3
         - sinon: 50%
+  note: Il appartient à l'utilisateur de bien vérifier que la somme des trois pourcentages renseignés vaut 100%.
         
 
 entreprise . activité . libérale réglementée:
