+++ mon-entreprise/source/pages/Simulateurs/AutoEntrepreneur.tsx
@@ -1,104 +1,22 @@
-import { batchUpdateSituation } from 'Actions/actions'
-import { Explicable } from 'Components/conversation/Explicable'
-import { Condition, WhenAlreadyDefined } from 'Components/EngineValue'
+import ChiffreAffairesActivitéMixte from 'Components/ChiffreAffairesActivitéMixte'
+import { WhenAlreadyDefined } from 'Components/EngineValue'
 import PeriodSwitch from 'Components/PeriodSwitch'
 import SimulateurWarning from 'Components/SimulateurWarning'
 import Simulation from 'Components/Simulation'
 import { SimulationGoal, SimulationGoals } from 'Components/SimulationGoals'
 import StackedBarChart from 'Components/StackedBarChart'
 import { ThemeColorsContext } from 'Components/utils/colors'
-import { useEngine } from 'Components/utils/EngineContext'
-import { Markdown } from 'Components/utils/markdown'
-import { serializeEvaluation } from 'publicodes'
-import { default as React, useCallback, useContext } from 'react'
+import { default as React, useContext } from 'react'
 import { Trans, useTranslation } from 'react-i18next'
-import { useDispatch, useSelector } from 'react-redux'
-import { situationSelector } from 'Selectors/simulationSelectors'
-
-const proportions = {
-	'entreprise . activité . mixte . proportions . service BIC':
-		"entreprise . chiffre d'affaires . service BIC",
-	'entreprise . activité . mixte . proportions . service BNC':
-		"entreprise . chiffre d'affaires . service BNC",
-	'entreprise . activité . mixte . proportions . vente restauration hébergement':
-		"entreprise . chiffre d'affaires . vente restauration hébergement",
-} as const
-function useAdjustProportions(): () => void {
-	const engine = useEngine()
-	const dispatch = useDispatch()
-
-	return useCallback(() => {
-		const nouveauCA = serializeEvaluation(
-			engine.evaluate({
-				somme: Object.values(proportions)
-					.map((chiffreAffaire) =>
-						serializeEvaluation(engine.evaluate(chiffreAffaire))
-					)
-					.filter(Boolean),
-			})
-		)
-		if (nouveauCA === '0€/an') {
-			return // Avoid division by 0
-		}
-		const situation = Object.entries(proportions).reduce(
-			(acc, [proportionName, valueName]) => {
-				const value = serializeEvaluation(
-					engine.evaluate({ valeur: valueName, 'par défaut': '0€/an' })
-				)
-				const newProportion = serializeEvaluation(
-					engine.evaluate({
-						valeur: `${value} / ${nouveauCA}`,
-						unité: '%',
-					})
-				)
-
-				return { ...acc, [proportionName]: newProportion }
-			},
-			{ "dirigeant . auto-entrepreneur . chiffre d'affaires": nouveauCA }
-		)
-		dispatch(batchUpdateSituation(situation))
-	}, [engine, dispatch])
-}
 
 export default function AutoEntrepreneur() {
-	const adjustProportions = useAdjustProportions()
-	const activitéMixte =
-		useEngine().evaluate('entreprise . activité . mixte').nodeValue === true
-
 	return (
 		<>
 			<SimulateurWarning simulateur="auto-entrepreneur" />
 			<Simulation explanations={<Explanation />}>
 				<PeriodSwitch />
 				<SimulationGoals className="plain">
-					<SimulationGoal
-						appear={false}
-						editable={!activitéMixte}
-						dottedName="dirigeant . auto-entrepreneur . chiffre d'affaires"
-					/>
-
-					<Condition expression="entreprise . activité . mixte">
-						<li className="small-target">
-							<ul>
-								<SimulationGoal
-									onUpdateSituation={adjustProportions}
-									dottedName="entreprise . chiffre d'affaires . vente restauration hébergement"
-								/>
-								<SimulationGoal
-									onUpdateSituation={adjustProportions}
-									dottedName="entreprise . chiffre d'affaires . service BIC"
-								/>
-								<SimulationGoal
-									onUpdateSituation={adjustProportions}
-									dottedName="entreprise . chiffre d'affaires . service BNC"
-								/>
-							</ul>
-						</li>
-					</Condition>
-					<ActivitéMixte
-						key={'' + activitéMixte}
-						defaultChecked={activitéMixte}
-					/>
+					<ChiffreAffairesActivitéMixte dottedName="dirigeant . auto-entrepreneur . chiffre d'affaires" />
 					<SimulationGoal
 						small
 						editable={false}
@@ -116,52 +34,6 @@ export default function AutoEntrepreneur() {
 				</SimulationGoals>
 			</Simulation>
 		</>
-	)
-}
-
-function ActivitéMixte({ defaultChecked }: { defaultChecked: boolean }) {
-	const dispatch = useDispatch()
-	const situation = useSelector(situationSelector)
-	const rule = useEngine().getRule('entreprise . activité . mixte')
-
-	const onMixteChecked = useCallback(
-		(checked: boolean) => {
-			dispatch(
-				batchUpdateSituation(
-					Object.values(proportions).reduce(
-						(acc, dottedName) => ({ ...acc, [dottedName]: undefined }),
-						{ 'entreprise . activité . mixte': checked ? 'oui' : 'non' }
-					)
-				)
-			)
-		},
-		[dispatch, situation]
-	)
-	return (
-		<li
-			className="small-target"
-			css={`
-				margin-top: -1rem;
-			`}
-		>
-			<label
-				css={`
-					display: flex;
-					align-items: center;
-					justify-content: flex-end;
-				`}
-			>
-				<input
-					type="checkbox"
-					defaultChecked={defaultChecked}
-					onChange={(evt) => onMixteChecked(evt.target.checked)}
-				/>
-				&nbsp; Activité mixte
-				<Explicable>
-					<Markdown source={`## ${rule.title}\n ${rule.rawNode.description}`} />
-				</Explicable>
-			</label>
-		</li>
 	)
 }
