+++ mon-entreprise/source/pages/Simulateurs/configs/auto-entrepreneur.yaml
@@ -22,9 +22,7 @@ questions:
   liste noire:
     - entreprise . charges
     - entreprise . chiffre d'affaires
-    - entreprise . chiffre d'affaires . vente restauration hébergement
-    - entreprise . chiffre d'affaires . service BIC
-    - entreprise . chiffre d'affaires . service BNC
+
 unité par défaut: €/an
 situation:
   entreprise . activité . mixte: non
